requirements.txt
twitter-api-client
python_dotenv
certifi
requests>=2.26.0